src/api-engine/src/requirements.txt
Django<=2.1.4
uwsgi<=2.0.17.1
djangorestframework<=3.9.0
django-filter<=1.1.0
drf-yasg<=1.11.1
flex<=6.13.2
swagger_spec_validator<=2.4.1